Cleanup: Extract some buildscript logic to prepare for expressions.
