+++ mixin-versions/build.gradle.kts
@@ -8,6 +8,7 @@ dependencies {
 
 subprojects {
     dependencies {
+        compileOnly(asm())
         compileOnly(rootProject)
     }
 }
