+++ mixin-versions/v0_8_3/build.gradle.kts
@@ -3,11 +3,6 @@ plugins {
 }
 
 dependencies {
+    compileOnly(mixin("0.8.3"))
     api(parent!!.project("v0_8"))
-}
-
-configurations.all {
-    resolutionStrategy {
-        force("org.spongepowered:mixin:0.8.3")
-    }
 }
