+++ mixin-versions/v0_8_4/build.gradle.kts
@@ -3,11 +3,6 @@ plugins {
 }
 
 dependencies {
+    compileOnly(mixin("0.8.4"))
     api(parent!!.project("v0_8_3"))
-}
-
-configurations.all {
-    resolutionStrategy {
-        force("org.spongepowered:mixin:0.8.4")
-    }
 }
